refactor: (컴포넌트 리팩토링) 도메인 관리를 RestaurantList 컴포넌트에서 담당하도록 수정 및 Filter 컴포넌트 로직 수정
