+++ src/componentsTmp/Filter.js
@@ -24,21 +24,19 @@ class Filter {
     this.$target.insertAdjacentHTML("beforeend", this.template());
   }
 
-  onChangeSelectBox(restaurantList, restaurantListManager) {
+  onChangeSelectBox(restaurantList) {
     this.$target.querySelector("#category-filter").addEventListener("change", (event) => {
       const selectedCategory = event.target.value;
       const selectedSortingWay = this.$target.querySelector("#sorting-filter").value;
 
-      const filteredList = restaurantListManager.getRestaurantList(selectedSortingWay, selectedCategory);
-      restaurantList.renderList(filteredList);
+      restaurantList.renderFilteredList(selectedCategory, selectedSortingWay);
     });
 
     this.$target.querySelector("#sorting-filter").addEventListener("change", (event) => {
       const selectedSortingWay = event.target.value;
       const selectedCategory = this.$target.querySelector("#category-filter").value;
 
-      const filteredList = restaurantListManager.getRestaurantList(selectedSortingWay, selectedCategory);
-      restaurantList.renderList(filteredList);
+      restaurantList.renderFilteredList(selectedCategory, selectedSortingWay);
     });
   }
 }
